Move code from cli_4666 to cli
